Version definitiva 2.1 de la aplicacion

+++ commands/clear_cmd/clear.py
@@ -30,19 +30,25 @@ def clear(args:list=[], options:dict={}, flags:list=[], nested_cmds:dict={}):
     try:
         clear_logger.info(" Eliminando Bases de Datos...")
         dbs = dbc.list_dbs(only_app_dbs=True)
-        for db in dbs:
-            dbc.drop_db(db)
-            clear_logger.info(f"'{db}' ha sido eliminada")
+        if len(dbs) > 0:
+            for db in dbs:
+                dbc.drop_db(db)
+                clear_logger.info(f"'{db}' ha sido eliminada")
+        else:
+            clear_logger.warning(' No existen bases de datos de la aplicacion')
         clear_logger.info(" Eliminando registro...")
         # mongoapp register clear
         clear_register(flags=['-y'])
         clear_logger.info(" Eliminando migraciones...")
-        os.remove('db.sqlite3')
+        if os.path.exists('db.sqlite3'):
+            os.remove('db.sqlite3')
+        else:
+            clear_logger.warning(' No existen migraciones actualmente')
     except Exception as err:
         msg = f" Error al vaciar la aplicacion\n     ERR MSG: {err}"
         msg += "\n -> Revisa que el servidor no este en funcionamiento"
         clear_logger.error(msg)
     else:
-        clear_logger.info(" Aplcacion vaciada con exito")
+        clear_logger.info(" La aplicacion esta vacia")
